refactor: extracting translateHref to a pure function
