+++ web/src/infra/i18n/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './href';
